refactor(mcp_server): extract write-dispatch and audit pipeline into write_dispatch.rs
Moves dispatch_write_tool, audit_write, audit_write_rejection,
audit_write_rejection_without_config, is_write_tool, summarize_write_args,
and related helpers out of the 2758-line tools.rs into a focused sibling
module (mcp_server/write_dispatch.rs). tools.rs delegates to write_dispatch::
at call sites; public behavior is unchanged. Adds 8 new targeted tests
covering the dispatch error path, is_write_tool, now_ms, first_chars, and
summarize_write_args edge cases.

Closes #2604
